Install improvements
