+++ lib/karafka/web/ui/models/health.rb
@@ -34,7 +34,7 @@ module Karafka
                 end
               end
 
-              raise Web::Errors::Ui::NotFoundError if stats.empty?
+             # raise Web::Errors::Ui::NotFoundError if stats.empty?
 
               stats
             end
