+++ lib/karafka/web/ui/views/shared/exceptions/not_found.erb
@@ -21,8 +21,7 @@
 
         <ul class="mb-5 text-start">
           <li>All the topics required by Karafka Web exist</li>
-          <li>Use <code>Karafka::Web.bootstrap_topics!</code> to create missing topics</li>
-          <li>You have started the <code>karafka server</code> at least once</li>
+          <li>You have used <code>Karafka::Web.bootstrap!</code> to initialize the Web UI</li>
           <li>You have a working connection with your Kafka cluster</li>
           <li>The resource you requested exists</li>
         </ul>
